move profile to this own app
