+++ oc_lettings_site/admin.py
@@ -1,6 +1,1 @@
 from django.contrib import admin
-
-from .models import Profile
-
-
-admin.site.register(Profile)
