+++ profiles/models/__init__.py
@@ -1,0 +1,1 @@
+from .models_profile import *
