+++ profiles/tests/__init__.py
@@ -1,0 +1,1 @@
+from .tests_lettings import *
